# Avoid inserting duplicate records in typerefs (fixes #61)