First naive attempt at solving https://github.com/wz1000/HieDb/issues/61.

Here's a quick comparison of this PR with master:

1. I built current master of haskell-language-server (to get bunch of .hie files to index):
`cabal clean && cabal build all --ghc-options=-fwrite-ide-info`

2. I indexed that directory with hiedb binary:
```bash
# build hiedb with optimizations:
cabal build --ghc-options=-O2
# find hiedb exe in cabal build dir and use it to index build artifacts in hls repo:
dist-newstyle/build/x86_64-linux/ghc-9.4.8/hiedb-0.5.0.1/x/hiedb/build/hiedb/hiedb -- -D tmp.db index ~/Devel/github.com/haskell/haskell-language-server/
```

- master: 
    - indexing stats: 303 indexed, 0 skipped in 1m13s + 0.07s gc
    - maximum residency: 68 908 kb
    - resulting sqlite size: 218MB
```bash
$ sqlite3 tmp.db "select count(*) from typerefs; select count(*) from (select distinct * from typerefs)"
342879 <<< ~3x redundancy
118819
```

- this PR: 
    - indexing stats: 303 indexed, 0 skipped in 56.61s + 0.07s gc <<< 1.3x faster
    - maximum residency: 69 924 kb
    - resulting sqlite size: 117MB <<< ~50% size reduction
    - no duplicates in sqlite:
```bash
$ sqlite3 tmp.db "select count(*) from typerefs; select count(*) from (select distinct * from typerefs)"
118819 <<< no redundancy
118819
```

You can see that difference with hls codebase is not that significant.
But the difference is much more significant with our work codebase which has much more deriving of stuff:

- master:
    - indexing stats: 208 indexed, 0 skipped in 1m39s + 0.04s gc
    - maximum residency: 68 900 kb
    - sqlite size: 771 MB
```bash
$ sqlite3 tmp.db "select count(*) from typerefs; select count(*) from (select distinct * from typerefs)"
2398458  <<< ~13x redundancy
189136
```
- this PR:
    - indexing stats: 208 indexed, 0 skipped in 41.16s + 0.05s gc  <<< 2.4x faster!
    - maximum residency: 55 232 kb  <<< wut? Even residency is smaller? Probably fluke
    - sqlite size: 102 MB   <<< size reduced by 87%
```bash
$ sqlite3 tmp.db "select count(*) from typerefs; select count(*) from (select distinct * from typerefs)"
189136 <<< no redundancy
189136
```