@wz1000 2024-02-08
I think we should add the UNIQUE constraint anyway so we can detect violations of this property.