@jhrcek 2024-02-08
> I think we should add the UNIQUE constraint anyway so we can detect violations of this property.

Would this require schema version bump?