@jhrcek 2024-02-08
Turns out when switching to int map/set I forgot to flip the boolean condition:
`when (Set.notMember ...)` ->  `unless (ISet.member ...`, that's why it was apparently faster.
Fixed in the last commit.