@wz1000 2024-02-08
Bump the schema version and this should be good to go I think