@jhrcek 2024-02-10
Hey @wz1000 
this should be ready to merge.
It would be also great to get this on hackage/use that in hls.
I can prepare a version bump PR if you want (and then push the update to hackage as I believe I'm still among the mainainers of this package).